pyproject.toml
[project]
name = "holobot"
version = "0.1.0"
description = "Conversational AI agent for the Holobox virtual human PoC — Tilburg public library"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "openai>=1.60",
    "elevenlabs>=1.20",
    "sounddevice>=0.5",
    "numpy>=2.0",
    "websockets>=14.0",
    "torch>=2.5",
    "torchaudio>=2.5",
]

[project.scripts]
holobot = "holobot.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/holobot"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
